Consolidate all MTP unit tests into a single file.
